// src/modules/qr-codes/dto/update-qr-code.dto.ts
export class UpdateQRCodeDto {
  type?: string;
  content?: string;
  ssid?: string;
  password?: string;
  link?: string;
  logoId?: number;
  backgroundColor?: string;
  padding?: number;
  logoBackgroundColor?: string;
  logoBorderRadius?: number;
  logoPadding?: number;
  qrCodeBorderRadius?: number;
}